+++ src/components/Genre.tsx
@@ -1,6 +1,7 @@
-import { useEffect, useState } from 'react'
-import { GenreData, ValueData } from '../Pages/Movies';
+import { useEffect } from 'react'
+import { ValueData } from '../Pages/Movies';
 import { APIKEY } from '../Services/api-client';
+import useGenre from '../hooks/useGenre';
 
 interface Props{
     type: string;
@@ -10,8 +11,7 @@ interface Props{
 
 
 const Genre = ({  type, value, setValue }: Props) => {
-  const [genre, setGenre] = useState<GenreData[]>([]);
-  const [page, setPage] = useState(1);
+  const {page,  genre, setGenre, CategoryAdd, CategoryRemove}  = useGenre({setValue, value});
   useEffect(() => {
     const fetchGenre = async () => {
       const data = await fetch(
@@ -20,25 +20,12 @@ const Genre = ({  type, value, setValue }: Props) => {
       const { genres } = await data.json();
       console.log(genres);
       setGenre(genres);
-      
-      
     };
     fetchGenre();
   }, [setGenre, type]);
   
   console.log(page);
-  const CategoryAdd = (genres: ValueData) => {
-    setValue([...value, genres]);
-    setGenre(genre.filter((g)=> g.id !== genres.id));
-    setPage(1)
-  }
 
-  const CategoryRemove = (genres: ValueData) => {
-    setValue(value.filter((g)=> g.id !== genres.id));
-    setGenre([...genre, genres]);
-    setPage(1)
-  }
-  
   return (
     <>
       <div className="genre-container">
